Improve dry-ness of base dockerfile.j2
The delorean repositories no longer have a separate location for
openvswitch.  Now openvswitch is located in delorean-deps.repo
and the rest of delorean master is located in the delorean.repo
file.  These files can be installed for both RDO and SOURCE, but
not for RHOS.  This patch uses the install_metatype to make a
determination as to when to install these two repos.  In the
process, we can remove the source RPM installation.

Change-Id: Ieedddd9d7ee234b6acdb03f7043d57c18e024951
Closes-Bug: #1508326