Use default disk image size for centos/7
Latest centos/7 image from atlas now comes with 40 GB disk size which
is equal to the virtual_size we specify in the Vagrantfile.  Libvirt
provider doesn't like it and produces an ugly stacktrace.

The new default value now sufficient run kolla properly and we can
remove the code that artificially increased the disk size.

Change-Id: I7a645543cfcc027b4a3e0d02d8e7f436d8410ec5
Backport: Liberty
Closes-Bug: #1511265