Remove unused setup_fedora.sh
The Fedora code is no longer used and can be removed. Should it be needed in the
future the code is very similiar to CentOS. As is it will just become out of
date as the code is never excersied.

Change-Id: I7df832e5b0830ac8b4507f000ed8ed6e43d39463
Partially-Implements: blueprint multinode-gate
(cherry picked from commit e4b84fbeacb892a47121541177af83b0117f91d7)