Add registry information to quickstart.rst
We get an awful lot of questions about how to operate with a registry
on the IRC channel and the obvious way to fix that is to document it.

I don't know what to document about Ubuntu but if someone leaves the
appropriate commands in the review I'll be happy to update it to include
the correct operations to make the registry run on Ubuntu.  Another option
is perhaps we can get Sam to write those docs once this hits the repo since
the structure will be mostly in place.

TrivialFix

Change-Id: Ib88abbaf9bd6bcabddae994157d9288aab8be2bf