Pin Liberty tarballs to tagged versions.
As discussed in Tokyo during the Kolla upgrade session [1], DB schema
changes can happen for minor releases which can be disruptive. In order
to prevent undesired DB migration, the Kolla developers decided to stop
tracking stable-liberty packages and instead stick to tagged versions.
If the new version require a DB migration, we'll write an Ansible
playbook to perform the upgrade.

[1] https://etherpad.openstack.org/p/kolla-mitaka-upgrade

Partially-Implements: blueprint upgrade-strategy

Change-Id: Iebef8a2090a5543afda5a4c90e75055c2d1f0e63