Add a new tool to help find the latest versions
As we will be switching to pinning versions in Kolla this tool will
help remove the 'manual' part of that equation by looking up the
versions we want.

The idea would be to eventually turn this into an autoproposal bot so
we only ever need to approve changes.

More work can be done here like automatically updating the
kolla-build.conf

Co-Authored-By: Martin André <m.andre@redhat.com>
Partially-Implements: blueprint upgrade-kolla

Change-Id: I052159850b2579ca6db79be71672be382be8a3a3