Adjust CMD for data and kolla-ansible
The CMD command for data causes docker to launch it with /bin/sh which
is ugly when viewing the containers statuses. Additionally for
kolla_ansible we were not specifying the CMD at all, forcing us to use
the 'command' param to set the sleep command.

This patch brings the kolla_ansible launch inline with the rest of the
project.

TrivialFix

Change-Id: Ic87a89921674802c15596756526e07f52a782f32