Fix broken nova-compute/nova-libvirt
Currently the state of master has nova_compute unable to talk to
nova-libvirt. There have been some efforts to fix this, but they have
failed for various reasons. This puts the connection type for libvirt
from a socket to connecting to the already-open tcp connection thus
restoring the ability for nova-compute to talk to nova-libvirt and
function.

Change-Id: I5a6c39bb16159a2924a93a831e3c53cedd0ca7a1
Partially-Implements: blueprint drop-root