Standardize bashate and add linters
The linters gate combindes pep8 and bashate (and then any other
linters we want). This changes bashate to be a sript rather than doing
a script inline. This matches the way other projects (and
project-config themselves) uses bashate [1].

Once a change upstream merges we can remove the redundant pep8 and
bashate gates.

[1] https://github.com/openstack-infra/project-config/blob/master/tox.ini#L11
https://github.com/openstack-infra/project-config/blob/master/tools/run-bashate.sh

Related-Id: I2afa7931bf3e18798b95a8af33ccf5ef761c4ad5

TrivialFix

Change-Id: I2ea2c1156e1b4de008f19e333e2740795dc406de