Revert "Capture the log in default"
This reverts commit b67c20f2db9da48fe5d3de1f5c5fb1887be1f20b.

This change hides the LOG output from the build process in the
console.html. While the build still will succeed/fail it will not update
the log while it is running. It will only dump the output of what its
doing if it fails. This is not ideal for our gates at this time since we
have seen the even the running and successful gates sometimes do not do
what we designed the gate to do (not building the containers we want,
not running the tests we want (recently happened with bashate), etc)

Change-Id: Ida58eef41ee8874e7577fcc24174fed4808f46f1