Use rdo magnum instead of sdake's copr for magnum
Steve is tired of maintaining a copr for Magnum.  People bug him
all the time to update the rpm for RDO.  The RDO community
has offered to take on the maintainance of the Magnum RPM. As this
RPM won't be in current-passed-ci for some time so it needs
to be pulled from current repo for the foreseeable future,
possibly nearing the release of Mitaka.

Change-Id: I9cfb02ab828251ef5bf40ca236f18b5f0f715e34
Closes-Bug: #1539325