Make kolla-toolbox container build on CentOS bin
Ansible installs pytz from pip which is newer then then pytz in
RDO repositories.  As part of this process, it removes pytz
which corrupts the python installation in some way.  This causes
further pip usage to produce a backtrace causing the kolla-toolbox
container not to build.

Mixing pip and RPM is not good.

Change-Id: I413468641bad5476cfcba61cb0e2189c31f3d682
Closes-Bug: #1545215