Make libvirt reestablish connection with qemu
Libvirt stores some information in /run at runtime that is needed to
automatically reestablish a connection with the VM when a new
container is created. Without this information a long (and manual)
process is needed to redefine the running vms and reattach to the
running qemu process.

This mountpoint was removed as "unneeded" in the past, but it does
exist in Liberty branch enabling a no-vm-downtime upgrade.

TrivialFix

Change-Id: I2eb31c602d8d17cbd6a8e405daf4123070794843