Install keystone-dist-paste.ini in /etc/keystone
The CentOS packages install a default version of the paste file
but don't install them in the /etc directory.  Kolla doesn't
configure the paste files, so this file never gets setup properly.

With the recent change in Keystone around the default admin_token
being changed, this triggered Keystone to return a 401 Auth error.

A follow-on patch from Sam edits this file, and that patch breaks
the build without this dependent patch.

TrivialFix

Change-Id: Ib1568e186bdd7d19b7e5af151388197755902488