Fix ubuntu marathon container
When we install mesos if it(or some deps) doesn't find java it installs
gcj(java5) which generate cacerts, so later in marathon we have
situation that openjdk8 can't access(discover as broken) to it because
of different format. Additionaly this third-party openjdk8 has in
dependencies ca-certificates-java which need java to generate cacerts
file.

Change-Id: Ic4467529a842ffc4bdd99ac1e23abbdee3a82422
Closes-bug: #1552626