Fix inconsistancy in globals.yml
The parameter values in global.yml were inconsist,
for some variable default values are shown while for
others it's not.

From user point of view it is important to know
the default values of the parameters and the globals.yml is
the file where user is supposed to look for config variable,
for sure a user do not want to look kolla/ansible/group_var/all.yml
file just for checking default values. So it is better to show all
default values in global.yml

This patch will solve this issue.

TrivialFix

Change-Id: I991fc5e1d4ed48d106da002a0f18a2b31525a482