Add reno support
For proper release note generation on the releases.openstack.org
website, this change will need to be backported into mitaka and
liberty as well.  Then a special one-file reno file will have to
be added for mitaka describing the features and the same story
for liberty.

See log of the discussion here:
http://eavesdrop.openstack.org/irclogs/%23openstack-release/%23openstack-release.2016-03-23.log.html#t2016-03-23T18:00:05

Change-Id: I3c298ae08ec4645f384ea5829cb0851fe82f4557
Partially-Implements: blueprint add-reno