Revert "Fix consistently failing CentOS gate"
This reverts commit 17e0de6aa6c32c5adebeffc90c96f15285877cf5.

This change should have never merged. There is something else going on
here and the bug doesnt fully invetigate the issue. Using a mix of
system and pip packages is not acceptable and can cause problems.
In this case the packaged mysql has a KNOWN issue that we fixed in
liberty by switching to the pip installed package.

Change-Id: Ie10116010156c13742a0a5d37c440c713e4698e5