Workaround ansible bug related to delegate_to
Currently the delegate_to doesnt happen and the neutron role creation is
attempted once on the first server and is skipped. The re-ordering of hosts in
site.yml seems to make the first host to be one inside neutron-server group
yielding the expected results. This patch needs to be re-visited as soon as a
version of ansible is chosen that fixes the issues with delegate_to

Co-Authored-By: Steven Dake <stdake@cisco.com>
Co-Authored-By: Vikram Hosakote <vhosakot@cisco.com>
Co-Authored-By: Nate Potter <nathaniel.potter@intel.com>
Co-Authored-By: Ganesh Mahalingam <ganesh.mahalingam@intel.com>
Change-Id: Ia712b323aa9d750d470a11ee899ab1b3054a903f
Partial-Bug: #1546789