Fix IP resolve check for RabbitMQ on AIO
On AIO installation we cannot assume that the public IP address
will be the first entry in "getent ahostsv4" result, because
it may be also a localhost address. To make this check positive
in AIO, we should look for the public IP in the whole output.

Change-Id: I1da7b95d7f00c7f87ff68ead46bf55fdea812599
Closes-Bug: 1564564
(cherry picked from commit 5aa2dcce3089d08a46175513fecb55ce91004769)