Increase max scheduling attempts to 10
At high scale, such as 64 nodes with 13TB ram and 2600 cores, nova
seems to struggle when scheduling 100+ VMs at the same time.  The
issue is unrelated to the database, as the error printed indicates
the max_scheduling_attempts have been reached.  Increase that value
to something more fitting of a 100 node cluster.

Change-Id: I8982d77c7c66db8f7c95b9fd73f58ceb66dbd723
Closes-Bug: #1563664