Increase max pool size so conductor doesn't implode
When horizon is used to launch 2000 VMs, nova-conductor is very
busy making database connections.  All 55 database connections are
in use, resulting in an inability to garbage collect database
connections.  Instead raise the max pool to 50 which will allow
50 concurrent database connections and the max overflow to 1000
which permits the database connections to finish the job at
large nodecount scales.

Closes-Bug: #1565105

Change-Id: I26dc2f7fda8760197888a1d61fbc45dfada2dd06